tidies up db queries
